refactor `Hints` type
