+++ modinfo/Cargo.toml
@@ -6,8 +6,7 @@ edition = "2021"
 license = "GPL-2.0-or-later"
 
 [dependencies]
-macros = { workspace = true }
+game = { workspace = true }
 log = { workspace = true }
-rand = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
